+++ test/macula_record_advertisement_tests.erl
@@ -199,7 +199,7 @@ delegation_bundle(Procedure, Overrides) ->
     Del = macula_record:sign(macula_record:procedure_delegation(
                                  macula_node_keys:key_id(DelSigner),
                                  maps:get(del_advertiser, Overrides, macula_node_keys:key_id(A)),
-                                 #{ttl_ms => maps:get(del_ttl, Overrides, 6 * ?HOUR)}), DelSigner),
+                                 #{ttl_ms => maps:get(del_ttl, Overrides, 30 * ?MINUTE)}), DelSigner),
     DirBytes = corrupt(maps:get(corrupt_dir, Overrides, false), macula_record:encode(OrgDir)),
     Authorization = #{org_directory => DirBytes, procedure_delegation => macula_record:encode(Del)},
     Adv = macula_record:sign(macula_record:procedure_advertisement(